requirements.txt
gevent==23.9.1
ipdb==0.13.9
pika==1.3.1
psutil==5.9.3
psycopg2-binary==2.9.5
pymysql==1.0.2
python-dateutil==2.8.2
responses==0.9.0
sortedcontainers==2.4.0
requests==2.31.0
urllib3==1.26.12
